refactor(target allocator): move telemetry setup to telemetry.go
