+++ cmd/otel-allocator/main.go
@@ -9,24 +9,15 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"regexp"
-	"strings"
 	"syscall"
 
 	"github.com/oklog/run"
 	monitoringclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/discovery"
-	prometheusbridge "go.opentelemetry.io/contrib/bridges/prometheus"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
-	otelprom "go.opentelemetry.io/otel/exporters/prometheus"
 	"go.opentelemetry.io/otel/metric"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.27.0"
 	"k8s.io/client-go/kubernetes"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -84,39 +75,15 @@ func main() {
 		os.Exit(1)
 	}
 
-	// The OTel SDK metrics are exported to Prometheus through a dedicated registry
-	// rather than the default one. This keeps them separate from the metrics that are
-	// registered directly on the default Prometheus registry (Prometheus service
-	// discovery internals, Go runtime and process collectors). That separation lets the
-	// OTLP reader below pull those Prometheus-only metrics via the Prometheus bridge
-	// without double-counting the SDK metrics, which it already collects natively.
-	sdkRegistry := prometheus.NewRegistry()
-	metricExporter, promErr := otelprom.New(otelprom.WithRegisterer(sdkRegistry))
-	if promErr != nil {
-		panic(promErr)
+	metricsGatherer, shutdownMeterProvider, mpErr := setupMeterProvider(ctx, cfg)
+	if mpErr != nil {
+		setupLog.Error(mpErr, "Unable to set up metrics")
+		os.Exit(1)
 	}
-	meterProviderOpts := []sdkmetric.Option{sdkmetric.WithReader(metricExporter)}
-
-	if cfg.Telemetry.Metrics.OTLP != nil {
-		otlpReader, otlpErr := newOTLPMetricReader(ctx, cfg.Telemetry.Metrics.OTLP)
-		if otlpErr != nil {
-			setupLog.Error(otlpErr, "Failed to create OTLP metric reader")
-			os.Exit(1)
-		}
-		telemetryResource, resErr := telemetryResource(ctx)
-		if resErr != nil {
-			setupLog.Error(resErr, "Failed to build telemetry resource")
-			os.Exit(1)
-		}
-		meterProviderOpts = append(meterProviderOpts, sdkmetric.WithReader(otlpReader), sdkmetric.WithResource(telemetryResource))
-	}
-
-	meterProvider := sdkmetric.NewMeterProvider(meterProviderOpts...)
-	otel.SetMeterProvider(meterProvider)
 	defer func() {
 		// Flush and close exporters (notably the OTLP reader) on shutdown so the final
 		// batch of metrics is delivered.
-		if shutdownErr := meterProvider.Shutdown(context.Background()); shutdownErr != nil {
+		if shutdownErr := shutdownMeterProvider(context.Background()); shutdownErr != nil {
 			setupLog.Error(shutdownErr, "Error shutting down meter provider")
 		}
 	}()
@@ -141,10 +108,10 @@ func main() {
 	if cfg.AllowInsecureAuthSecrets {
 		httpOptions = append(httpOptions, server.WithInsecureAuthSecrets())
 	}
-	// Serve /metrics from both the default registry (Prometheus service discovery, Go
-	// runtime and process collectors) and the dedicated SDK registry (OTel SDK metrics),
-	// preserving the full metric set that used to be exposed via the default registry alone.
-	httpOptions = append(httpOptions, server.WithMetricsGatherer(prometheus.Gatherers{prometheus.DefaultGatherer, sdkRegistry}))
+	// Serve /metrics from the union of the default registry (Prometheus service discovery,
+	// Go runtime and process collectors) and the dedicated SDK registry (OTel SDK metrics),
+	// as assembled by setupMeterProvider.
+	httpOptions = append(httpOptions, server.WithMetricsGatherer(metricsGatherer))
 	srv, serverErr := server.NewServer(log, allocator, cfg.ListenAddr, httpOptions...)
 	if serverErr != nil {
 		panic(serverErr)
@@ -336,137 +303,4 @@ func main() {
 		setupLog.Error(runErr, "run group exited")
 	}
 	setupLog.Info("Target allocator exited.")
-}
-
-// envRefPattern matches ${env:VAR} references, mirroring the OTel Collector's
-// environment-variable substitution syntax.
-var envRefPattern = regexp.MustCompile(`\$\{env:([a-zA-Z_][a-zA-Z0-9_]*)\}`)
-
-// expandEnvRefs replaces ${env:VAR} occurrences in s with the value of the
-// environment variable VAR (empty if unset). This lets sensitive header values
-// (e.g. an API token) be sourced from the pod environment / a Secret at runtime
-// rather than being written in plaintext into the Target Allocator ConfigMap.
-func expandEnvRefs(s string) string {
-	return envRefPattern.ReplaceAllStringFunc(s, func(match string) string {
-		name := envRefPattern.FindStringSubmatch(match)[1]
-		return os.Getenv(name)
-	})
-}
-
-// expandHeaders returns a copy of headers with ${env:VAR} references expanded in
-// every value.
-func expandHeaders(headers map[string]string) map[string]string {
-	if len(headers) == 0 {
-		return nil
-	}
-	expanded := make(map[string]string, len(headers))
-	for k, v := range headers {
-		expanded[k] = expandEnvRefs(v)
-	}
-	return expanded
-}
-
-// newOTLPMetricReader creates a PeriodicExportingMetricReader backed by either
-// an OTLP/gRPC or OTLP/HTTP exporter depending on cfg.Protocol.
-func newOTLPMetricReader(ctx context.Context, cfg *config.OTLPExporterConfig) (sdkmetric.Reader, error) {
-	temporality := temporalitySelector(cfg.Temporality)
-	readerOpts := periodicReaderOptions(cfg)
-	endpoint := expandEnvRefs(cfg.Endpoint)
-	headers := expandHeaders(cfg.Headers)
-	switch cfg.Protocol {
-	case "http":
-		// WithEndpointURL does not append /v1/metrics automatically. We do it here
-		// so the endpoint convention matches the OTel collector's otlphttp exporter
-		// (where users specify the base URL, not the full signal path).
-		// Skip if the caller already included the signal path.
-		endpoint = strings.TrimRight(endpoint, "/")
-		if !strings.HasSuffix(endpoint, "/v1/metrics") {
-			endpoint += "/v1/metrics"
-		}
-		opts := []otlpmetrichttp.Option{
-			otlpmetrichttp.WithEndpointURL(endpoint),
-			otlpmetrichttp.WithTemporalitySelector(temporality),
-		}
-		if cfg.Insecure {
-			opts = append(opts, otlpmetrichttp.WithInsecure())
-		}
-		if len(headers) > 0 {
-			opts = append(opts, otlpmetrichttp.WithHeaders(headers))
-		}
-		exp, err := otlpmetrichttp.New(ctx, opts...)
-		if err != nil {
-			return nil, err
-		}
-		return sdkmetric.NewPeriodicReader(exp, readerOpts...), nil
-	default: // "grpc"
-		opts := []otlpmetricgrpc.Option{
-			otlpmetricgrpc.WithTemporalitySelector(temporality),
-		}
-		// Accept both a full URL (with scheme, consistent with the HTTP protocol and
-		// the collector's otlp exporter) and a bare host:port. WithEndpointURL requires
-		// a scheme; WithEndpoint expects host:port.
-		if strings.Contains(endpoint, "://") {
-			opts = append(opts, otlpmetricgrpc.WithEndpointURL(endpoint))
-		} else {
-			opts = append(opts, otlpmetricgrpc.WithEndpoint(endpoint))
-		}
-		if cfg.Insecure {
-			opts = append(opts, otlpmetricgrpc.WithInsecure())
-		}
-		if len(headers) > 0 {
-			opts = append(opts, otlpmetricgrpc.WithHeaders(headers))
-		}
-		exp, err := otlpmetricgrpc.New(ctx, opts...)
-		if err != nil {
-			return nil, err
-		}
-		return sdkmetric.NewPeriodicReader(exp, readerOpts...), nil
-	}
-}
-
-// periodicReaderOptions builds PeriodicReaderOptions from the export interval and timeout config.
-func periodicReaderOptions(cfg *config.OTLPExporterConfig) []sdkmetric.PeriodicReaderOption {
-	opts := []sdkmetric.PeriodicReaderOption{
-		// Bridge the metrics registered directly on the default Prometheus registry
-		// (Prometheus service discovery internals, Go runtime and process collectors)
-		// into the OTLP export. Without this, those metrics would be visible on the
-		// Prometheus /metrics endpoint but missing from OTLP. The SDK's own metrics live
-		// on a separate registry and are collected natively by this reader, so bridging
-		// the default registry does not double-count them.
-		sdkmetric.WithProducer(prometheusbridge.NewMetricProducer()),
-	}
-	if cfg.ExportInterval > 0 {
-		opts = append(opts, sdkmetric.WithInterval(cfg.ExportInterval))
-	}
-	if cfg.Timeout > 0 {
-		opts = append(opts, sdkmetric.WithTimeout(cfg.Timeout))
-	}
-	return opts
-}
-
-// telemetryResource builds the OpenTelemetry resource attached to exported metrics.
-// It sets a default service.name and honors the standard OTEL_SERVICE_NAME /
-// OTEL_RESOURCE_ATTRIBUTES environment variables, so additional attributes (e.g. k8s.*)
-// can be injected without code changes.
-func telemetryResource(ctx context.Context) (*resource.Resource, error) {
-	return resource.New(ctx,
-		resource.WithTelemetrySDK(),
-		resource.WithAttributes(semconv.ServiceName("target-allocator")),
-		// Applied last so environment configuration wins on conflict.
-		resource.WithFromEnv(),
-	)
-}
-
-// temporalitySelector maps a config string to an SDK TemporalitySelector.
-// "delta" → all instruments delta; "lowmemory" → delta for counters/histograms,
-// cumulative for gauges; anything else → cumulative (SDK default).
-func temporalitySelector(t string) sdkmetric.TemporalitySelector {
-	switch t {
-	case "delta":
-		return sdkmetric.DeltaTemporalitySelector
-	case "lowmemory":
-		return sdkmetric.LowMemoryTemporalitySelector
-	default:
-		return sdkmetric.DefaultTemporalitySelector
-	}
 }
